Fix type conversion warnings in examples
Fix type conversion warnings in examples occurring when using
-Wconversion flag of gcc. The commit contains also
tests/rest_server_test.cc with the similar fix.

One of questionable change types was caused by
Http::Endpoint::options()::threads(int val), and was invoked with size_t
argument. Maybe it makes sense to modify the function argument type.

After this commit, there still remain a lot of -Wconversion warnings in
rapidjson library

+++ examples/http_server.cc
@@ -157,10 +157,10 @@ int main(int argc, char *argv[]) {
     int thr = 2;
 
     if (argc >= 2) {
-        port = std::stol(argv[1]);
+        port = static_cast<uint16_t>(std::stol(argv[1]));
 
         if (argc == 3)
-            thr = std::stol(argv[2]);
+            thr = std::stoi(argv[2]);
     }
 
     Address addr(Ipv4::any(), port);
